Refactor TV rotation: dynamic image loading, 30s duration, and Ken Burns animation
